webapp/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==2.1.3  # Compatible with Werkzeug<2.2.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pip==24.2
SQLAlchemy==1.4.46
typing_extensions==4.12.2
Werkzeug==2.1.2  # Compatible with moto
pytest==8.3.3
fastapi
requests
flask_bcrypt
flask_sqlalchemy
flask_httpauth
python-dotenv
pymysql
cryptography
boto3
watchtower
sendgrid
statsd
moto
